+++ src/main/java/com/hust/baseweb/entity/Person.java
@@ -1,4 +1,4 @@
-package com.hust.baseweb.model;
+package com.hust.baseweb.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
