+++ src/main/java/com/hust/baseweb/entity/SecurityPermission.java
@@ -1,4 +1,4 @@
-package com.hust.baseweb.model;
+package com.hust.baseweb.entity;
 
 import lombok.Getter;
 import lombok.Setter;
