+++ src/main/java/com/hust/baseweb/repo/PartyRepo.java
@@ -1,6 +1,6 @@
 package com.hust.baseweb.repo;
 
-import com.hust.baseweb.model.Party;
+import com.hust.baseweb.entity.Party;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
