+++ src/main/java/com/hust/baseweb/repo/PartyTypeRepo.java
@@ -1,6 +1,6 @@
 package com.hust.baseweb.repo;
 
-import com.hust.baseweb.model.PartyType;
+import com.hust.baseweb.entity.PartyType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
