+++ src/main/java/com/hust/baseweb/repo/PersonRepo.java
@@ -1,6 +1,6 @@
 package com.hust.baseweb.repo;
 
-import com.hust.baseweb.model.Person;
+import com.hust.baseweb.entity.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
