+++ src/main/java/com/hust/baseweb/repo/StatusRepo.java
@@ -1,6 +1,6 @@
 package com.hust.baseweb.repo;
 
-import com.hust.baseweb.model.Status;
+import com.hust.baseweb.entity.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
