+++ src/main/java/com/hust/baseweb/repo/UserLoginRepo.java
@@ -1,7 +1,7 @@
 package com.hust.baseweb.repo;
 
-import com.hust.baseweb.model.Party;
-import com.hust.baseweb.model.UserLogin;
+import com.hust.baseweb.entity.Party;
+import com.hust.baseweb.entity.UserLogin;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
